repo_arm_checker/requirements.txt
requests==2.31.0
langchain==0.0.267
langchain-google-genai==0.0.5
python-dotenv==1.0.0
